Gradle script changes, reorganized files and removed CardinalFinancialDemo files
