+++ build.gradle
@@ -13,9 +13,6 @@ plugins {
     // Apply the application plugin to add support for building a CLI application.
     id 'application'
 }
-ext {
-	javaMainClass = "com.cardinalfinancial.us.main.CardinalFinancialDemo"
-}
 repositories {
     // Use jcenter for resolving dependencies.
     // You can declare any Maven/Ivy/file repository here.
@@ -30,8 +27,12 @@ dependencies {
     testImplementation 'junit:junit:4.13'
 }
 
-application {
+/*application {
     // Define the main class for the application.
-    mainClassName = javaMainClass
     sourceSets.main.java.srcDirs = ['src']
+}*/
+sourceSets {
+       test {
+               java.srcDirs = ['src/main/', 'src/test/']
+       }
 }
