+++ src/main/java/com/cardinalfinancial/us/model/Checkout.java
